pyproject.toml
[tool.poetry]
name = "evolve_fb_ads_insights"
version = "2"
description = ""
authors = ["HM <hieumdd@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.9"
requests = "^2.26.0"
google-cloud-bigquery = "^2.28.1"
Jinja2 = "^3.0.2"
aiohttp = "^3.7.4"
google-cloud-pubsub = "^2.8.0"
google-cloud-tasks = "^2.6.0"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pytest = "^6.2.5"
pytest-timeout = "^2.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-s --durations=0"
timeout = 530
testpaths = [
    "test",
]
